requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu 
torchvision==0.9.1+cpu
Flask
gunicorn
werkzeug
numpy
pandas
Pillow
requests
matplotlib